requirements.txt
numpy==1.21.2
Pillow==8.4.0
pygame==2.1.2
RPi.GPIO==0.7.0
mido==1.2.10
json5==0.9.6
python-dateutil==2.8.2
functools==0.5
pyserial==3.5
PyYAML==5.4.1
RPi.GPIO==0.7.0
threading==0.1
tk==0.1.0
tkinter==0.1.0
py-midi==2.0.1
os==0.1.0
itertools==0.1.0
math==0.1.0
